// frontend/src/types/generated/SizeFilter.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Variable } from "./Variable";

export type SizeFilter =
  | {
      type: "NumChilds";
      child_name: string;
      min: number | null;
      max: number | null;
    }
  | { type: "BindingSetEqual"; child_names: Array<string> }
  | {
      type: "BindingSetProjectionEqual";
      child_name_with_var_name: Array<[string, Variable]>;
    }
  | {
      type: "NumChildsProj";
      child_name: string;
      var_name: Variable;
      min: number | null;
      max: number | null;
    }
  | { type: "AdvancedCEL"; cel: string };